Move note entity class from domain to data package
