+++ app/src/main/java/com/dogactnrvrdi/notesapp/data/source/local/NoteDao.kt
@@ -6,7 +6,7 @@ import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.dogactnrvrdi.notesapp.domain.model.Note
+import com.dogactnrvrdi.notesapp.data.model.Note
 import kotlinx.coroutines.flow.Flow
 
 @Dao
