+++ app/src/main/java/com/dogactnrvrdi/notesapp/data/source/local/NoteDatabase.kt
@@ -3,7 +3,7 @@ package com.dogactnrvrdi.notesapp.data.source.local
 import androidx.room.AutoMigration
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.dogactnrvrdi.notesapp.domain.model.Note
+import com.dogactnrvrdi.notesapp.data.model.Note
 
 @Database(
     entities = [Note::class],
