+++ app/src/main/java/com/dogactnrvrdi/notesapp/domain/use_case/AddNoteUseCase.kt
@@ -1,8 +1,8 @@
 package com.dogactnrvrdi.notesapp.domain.use_case
 
 import com.dogactnrvrdi.notesapp.R
-import com.dogactnrvrdi.notesapp.domain.model.InvalidNoteException
-import com.dogactnrvrdi.notesapp.domain.model.Note
+import com.dogactnrvrdi.notesapp.data.model.InvalidNoteException
+import com.dogactnrvrdi.notesapp.data.model.Note
 import com.dogactnrvrdi.notesapp.domain.repo.NoteRepository
 import kotlinx.coroutines.channels.Channel
 import javax.inject.Inject
