+++ app/src/main/java/com/dogactnrvrdi/notesapp/domain/use_case/GetNoteUseCase.kt
@@ -1,6 +1,6 @@
 package com.dogactnrvrdi.notesapp.domain.use_case
 
-import com.dogactnrvrdi.notesapp.domain.model.Note
+import com.dogactnrvrdi.notesapp.data.model.Note
 import com.dogactnrvrdi.notesapp.domain.repo.NoteRepository
 import javax.inject.Inject
