+++ app/src/main/java/com/dogactnrvrdi/notesapp/presentation/add_edit_note/AddEditNoteViewModel.kt
@@ -9,8 +9,8 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.dogactnrvrdi.notesapp.R
-import com.dogactnrvrdi.notesapp.domain.model.InvalidNoteException
-import com.dogactnrvrdi.notesapp.domain.model.Note
+import com.dogactnrvrdi.notesapp.data.model.InvalidNoteException
+import com.dogactnrvrdi.notesapp.data.model.Note
 import com.dogactnrvrdi.notesapp.domain.use_case.NoteUseCases
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
