+++ app/src/main/java/com/dogactnrvrdi/notesapp/presentation/notes/NotesState.kt
@@ -2,7 +2,7 @@ package com.dogactnrvrdi.notesapp.presentation.notes
 
 import com.dogactnrvrdi.notesapp.common.NoteOrder
 import com.dogactnrvrdi.notesapp.common.OrderType
-import com.dogactnrvrdi.notesapp.domain.model.Note
+import com.dogactnrvrdi.notesapp.data.model.Note
 
 data class NotesState(
     val notes: List<Note> = emptyList(),
